[step2] 뷰에서 카드를 표현하는 객체 CardExpression 정의
- https://github.com/next-step/kotlin-blackjack/pull/725#discussion_r1861727904
