+++ src/main/kotlin/blackjack/view/ResultView.kt
@@ -1,9 +1,11 @@
-package blackjack
+package blackjack.view
+
+import blackjack.BlackJackGameResult
 
 class ResultView {
     fun printBlackJackGameResult(blackJackGameResults: List<BlackJackGameResult>) {
         blackJackGameResults.forEach {
-            val displayCards = it.cards.joinToString(", ") { drawCard -> drawCard.drawCardString }
+            val displayCards = it.cards.joinToString(", ") { drawCard -> CardExpression.of(drawCard).value }
             println("${it.playerName}: $displayCards - 결과: ${it.totalValue}")
         }
     }
